Implement the rest. Next up; unit testing
