+++ module-src/vistas-client/src/main/java/com/craftmend/vistas/client/Vistas.java
@@ -1,13 +1,33 @@
 package com.craftmend.vistas.client;
 
+import com.craftmend.openaudiomc.OpenAudioMc;
 import com.craftmend.openaudiomc.api.enums.ModuleEvent;
 import com.craftmend.openaudiomc.api.interfaces.ExternalModule;
+import com.craftmend.openaudiomc.generic.commands.CommandService;
 import com.craftmend.openaudiomc.generic.environment.MagicValue;
+import com.craftmend.openaudiomc.generic.proxy.interfaces.UserHooks;
+import com.craftmend.openaudiomc.spigot.OpenAudioMcSpigot;
+import com.craftmend.vistas.client.client.ClientUserHooks;
+import com.craftmend.vistas.client.client.VistasRedisClient;
+import com.craftmend.vistas.client.commands.VistasEvalCommand;
+import com.craftmend.vistas.client.listeners.PlayerListener;
+import com.craftmend.vistas.client.redis.packets.ServerRegisterPacket;
+import com.craftmend.vistas.client.redis.packets.UserJoinPacket;
+import com.craftmend.vistas.client.server.networking.VistasRedisServer;
+import lombok.Getter;
+import org.bukkit.Bukkit;
+import org.bukkit.entity.Player;
 import org.bukkit.event.Listener;
+
+import java.util.UUID;
 
 public final class Vistas extends ExternalModule implements Listener {
 
+    @Getter private static Vistas instance;
+    @Getter private UUID serverId = UUID.randomUUID();
+
     public Vistas() {
+        instance = this;
         MagicValue.overWrite(MagicValue.NOTIFY_VOICECHAT_SLOT_DEPLETION, false);
         MagicValue.overWrite(MagicValue.FORCE_SERVER_NODE, true);
     }
@@ -29,7 +49,31 @@ public final class Vistas extends ExternalModule implements Listener {
 
     @Override
     public void on(ModuleEvent event) {
+        if (event == ModuleEvent.PLATFORM_LOADED) {
+            // finished startup
+            OpenAudioMc.getService(CommandService.class).registerSubCommand(new VistasEvalCommand());
+        }
 
+        if (event == ModuleEvent.SERVICES_LOADED) {
+            // mid boot
+            OpenAudioMc.getInstance().getServiceManager().loadServices(VistasRedisClient.class);
+            OpenAudioMc.getInstance().getServiceManager().registerDependency(UserHooks.class, new ClientUserHooks());
+            OpenAudioMcSpigot s = OpenAudioMcSpigot.getInstance();
+            s.getServer().getPluginManager().registerEvents(new PlayerListener(this), s);
+        }
+    }
+
+    public void registerSelf() {
+        OpenAudioMc.getService(VistasRedisClient.class).sendPacket(new ServerRegisterPacket(serverId));
+        for (Player onlinePlayer : Bukkit.getOnlinePlayers()) {
+            OpenAudioMc.getService(VistasRedisClient.class).sendPacket(
+                    new UserJoinPacket(
+                            onlinePlayer.getName(),
+                            onlinePlayer.getUniqueId(),
+                            serverId
+                    )
+            );
+        }
     }
 
 }
