--- module-src/vistas-client/src/main/java/com/craftmend/vistas/client/packets/InvokeUserPacket.java
@@ -1,19 +1,0 @@
-package com.craftmend.vistas.client.packets;
-
-import com.craftmend.openaudiomc.generic.networking.abstracts.AbstractPacketPayload;
-import com.craftmend.vistas.client.reflection.SerializedCall;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.util.UUID;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class InvokeUserPacket extends AbstractPacketPayload {
-
-    private SerializedCall call;
-    private UUID playerUuid;
-
-}
