--- module-src/vistas-client/src/main/java/com/craftmend/vistas/client/packets/WrappedProxyPacket.java
@@ -1,20 +1,0 @@
-package com.craftmend.vistas.client.packets;
-
-import com.craftmend.openaudiomc.generic.networking.abstracts.AbstractPacketPayload;
-import com.craftmend.openaudiomc.generic.proxy.messages.StandardPacket;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.util.UUID;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class WrappedProxyPacket extends AbstractPacketPayload {
-
-    private StandardPacket packet;
-    private UUID serverId;
-    private UUID playerId;
-
-}
