--- module-src/vistas-server/src/main/java/com/craftmend/vistas/server/users/MinecraftServer.java
@@ -1,56 +1,0 @@
-package com.craftmend.vistas.server.users;
-
-import com.craftmend.openaudiomc.OpenAudioMc;
-import com.craftmend.openaudiomc.generic.proxy.interfaces.UserHooks;
-import com.craftmend.openaudiomc.generic.proxy.messages.StandardPacket;
-import com.craftmend.openaudiomc.generic.proxy.models.ProxyNode;
-import com.craftmend.openaudiomc.generic.user.User;
-import com.craftmend.vistas.client.packets.WrappedProxyPacket;
-import com.craftmend.vistas.server.networking.VistasNetworkServer;
-import lombok.AllArgsConstructor;
-
-import java.util.Collection;
-import java.util.HashSet;
-import java.util.Set;
-import java.util.UUID;
-
-@AllArgsConstructor
-public class MinecraftServer implements ProxyNode {
-
-    private String name;
-    private UUID uuid;
-
-    @Override
-    public String getName() {
-        return name;
-    }
-
-    @Override
-    public Collection<User> getOnlineUsers() {
-        Set<User> users = new HashSet<>();
-        for (User onlineUser : OpenAudioMc.resolveDependency(UserHooks.class).getOnlineUsers()) {
-            if (onlineUser instanceof VistasServerUser) {
-                VistasServerUser du = (VistasServerUser) onlineUser;
-                if (du.getLastSeenServer() != null && du.getLastSeenServer().equals(uuid)) users.add(du);
-            }
-        }
-        return users;
-    }
-
-    public void sendPacket(User user, StandardPacket standardPacket) {
-        OpenAudioMc.getService(VistasNetworkServer.class).sendPacket(new WrappedProxyPacket(
-                standardPacket,
-                uuid,
-                user.getUniqueId()
-        ), uuid);
-    }
-
-    @Override
-    public void sendPacket(StandardPacket standardPacket) {
-        OpenAudioMc.getService(VistasNetworkServer.class).sendPacket(new WrappedProxyPacket(
-                standardPacket,
-                uuid,
-                null
-        ), uuid);
-    }
-}
